Rename camel case to snake case for convention
