+++ libs/langchain-google-cloud-sql-pg/src/index.ts
@@ -1,4 +1,4 @@
 export * from "./engine.js";
-export * from "./chatMessageHistory.js";
-export * from "./vectorStore.js";
+export * from "./chat_message_history.js";
+export * from "./vectorstore.js";
 export * from "./loader.js";
