+++ libs/langchain-google-cloud-sql-pg/src/tests/chat_message_history.ts
@@ -2,7 +2,7 @@ import { test } from "@jest/globals";
 import * as dotenv from "dotenv";
 import { AIMessage, BaseMessage, HumanMessage } from "@langchain/core/messages";
 import PostgresEngine from "../engine.js";
-import { PostgresChatMessageHistory } from "../chatMessageHistory.js";
+import { PostgresChatMessageHistory } from "../chat_message_history.js";
 
 dotenv.config();
